examples/src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
openai==1.3.0
azure-monitor-opentelemetry==1.2.0
opencensus-ext-azure==1.1.13
requests==2.31.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0